optimised build for linux

+++ build.py
@@ -29,7 +29,12 @@ if sys.platform == "darwin":
     shutil.copy(project_path + "Client/Resources/Structures.asset", project_path + "Output/MacOS/")
     shutil.rmtree(project_path + "Temp/")
 elif sys.platform == "linux":
-    os.system("sudo apt install libsdl2-dev libsdl2-image-dev libsdl2-ttf-dev")
+    lib_files = ["/usr/lib/x86_64-linux-gnu/libSDL2-2.0.so.0", "/usr/lib/x86_64-linux-gnu/libSDL2_image-2.0.so.0", "/usr/lib/x86_64-linux-gnu/libSDL2_image-2.0.so.0"]
+
+    for lib_file in lib_files:
+        if not os.path.exists(lib_file)
+            os.system("sudo apt install libsdl2-dev libsdl2-image-dev libsdl2-ttf-dev")
+            break
 
     createDir("Build/")
     os.system(f"cd {project_path}Build/ && cmake -DCMAKE_CXX_COMPILER=/usr/bin/clang++ .. && make -j$(nproc)")
@@ -42,9 +47,8 @@ elif sys.platform == "linux":
     shutil.rmtree(project_path + "Output/Linux/Terralistic/Resources/", ignore_errors=True)
     shutil.move(project_path + "Build/Resources/", project_path + "Output/Linux/Terralistic/")
 
-    shutil.copy("/usr/lib/x86_64-linux-gnu/libSDL2-2.0.so.0", project_path + "Output/Linux/Terralistic/")
-    shutil.copy("/usr/lib/x86_64-linux-gnu/libSDL2_image-2.0.so.0", project_path + "Output/Linux/Terralistic/")
-    shutil.copy("/usr/lib/x86_64-linux-gnu/libSDL2_image-2.0.so.0", project_path + "Output/Linux/Terralistic/")
+    for lib_file in lib_files:
+        shutil.copy(lib_file, project_path + "Output/Linux/Terralistic/")
 
     shutil.copy(project_path + "Build/Terralistic-server", project_path + "Output/Linux/Terralistic/")
     shutil.copy(project_path + "Build/Structures.asset", project_path + "Output/Linux/Terralistic/")
